+++ src/setupTests.ts
@@ -4,6 +4,7 @@ import '@testing-library/jest-dom/vitest';
 
 afterEach(() => {
   cleanup();
+  localStorage.clear();
 });
 
 class ResizeObserverStub {